+++ src/shape/cache.rs
@@ -3,7 +3,7 @@ use super::engine::EngineMetadata;
 use super::internal::var::Fvar;
 use crate::{charmap::CharmapProxy, metrics::MetricsProxy, FontRef};
 
-pub type Epoch = u64;
+pub type Epoch = usize;
 
 pub struct FontEntry {
     pub metrics: MetricsProxy,
@@ -27,7 +27,7 @@ impl FontEntry {
 
 pub struct FeatureEntry {
     pub epoch: Epoch,
-    pub id: u64,
+    pub id: usize,
     pub coords: Vec<i16>,
     pub tags: [u32; 4],
     pub store: FeatureStore,
@@ -55,7 +55,7 @@ impl FeatureCache {
 
     pub fn entry<'a>(
         &'a mut self,
-        id: u64,
+        id: usize,
         coords: &[i16],
         has_feature_vars: bool,
         tags: &[u32; 4],
@@ -77,7 +77,7 @@ impl FeatureCache {
 
     fn find_entry(
         &mut self,
-        id: u64,
+        id: usize,
         coords: &[i16],
         has_feature_vars: bool,
         tags: &[u32; 4],
